Temporary workaround for wireless suspend/resume.
Mark the wlan0 interface down on suspend and back up on resume to
prod connman into action.  The result isn't exactly what we want
as connman won't return to the previous ap if another, previously
visited ap, is found with higher signal strength.  Also this is
very heavyweight, the better behaviour is to simulate a beacon miss
on resume and try to reassociate to the same ap; then roam if that
fails--but getting that behaviour will take more work so add this
for now.
Review URL: http://chromereview.prom.corp.google.com/1175077

git-svn-id: svn://chrome-svn/chromeos/trunk@57 06c00378-0e64-4dae-be16-12b19f9950a1