wm: Fix a typo that could lead to crashes at startup.
I was overwriting the 'type' field of the client messages
that get sent to the root window at startup.  XSendEvent()
uses this to index into an array of function pointers, and
depending on the value of the atom that I was writing to the
wrong place, arbitrary event-sending handlers could be
invoked.  I was seeing segfaults in XInput code.

Review URL: http://chromereview.prom.corp.google.com/1173129

git-svn-id: svn://chrome-svn/chromeos/trunk@80 06c00378-0e64-4dae-be16-12b19f9950a1