wm: Move transient window code into ToplevelWindow.
The Window class wasn't the right place for it, and
tracking focus between a toplevel window and its
transients is much easier when they're all tracked
in a single place.  This also simplifies things in
that we don't pay attention to changes to the
transient-for hint after a window has been mapped,
and don't need to worry about weird stuff like
transient-for loops.

Review URL: http://chromereview.prom.corp.google.com/1177011

git-svn-id: svn://chrome-svn/chromeos/trunk@148 06c00378-0e64-4dae-be16-12b19f9950a1