init: scripts for starting X on our own, getting ready to boot to chrome
To use this stuff, look at /etc/init/ui.conf and heed the comments
there.  Basically, what this does is, instead of starting slim (which
currently starts X and stuff, it starts a control script which kicks
off X, waits for it, and then sets up the environment that Chrome
needs to get going and execs a small script as the "chronos" user.

Review URL: http://chromereview.prom.corp.google.com/1184001


git-svn-id: svn://chrome-svn/chromeos/trunk@183 06c00378-0e64-4dae-be16-12b19f9950a1