Autotest: install into stateful partition
Keep in mind that the stateful partition is kept around across system
updates, so we may want to set up something to "rm -rf
/mnt/stateful_partition/autotest" after an upgrade (perhaps when
re-install autotest on the new system?).

Also, for this test to run, the the root filesystem must be mounted
read-only. In practice, I seem to have to remount the filesystem
read-only before logging in via slim. At some point, we may want to
change the devkit to not mount the system read-write at boot
automatically.

Review URL: http://chromereview.prom.corp.google.com/1188049


git-svn-id: svn://chrome-svn/chromeos/trunk@251 06c00378-0e64-4dae-be16-12b19f9950a1