Overhaul microbenchmark so that it can be pulled in easily by any package
This follows the model set by gtest_main.a and adds a microbenchmark_main.a
which can be pulled in by any package to do quick microbenchmarking of any of
the functions in their tree.

This change also includes updates to minijail as an example usage of the microbenchmark code.

Last but not least, I fixed a copy and paste error in the copyright and a
missing comma in the output as well as updated some comments.

Review URL: http://codereview.chromium.org/492005