Add ARM cross-compiler to chroot.
This adds the ARM cross-compiler toolchain to the build chroot environment,
so we can start trying to compile things for ARM.

There's also a minor change to make_local_repo.sh and make_chroot.sh to
allow a second repository source. This is needed only for bootstrapping this
change (and others like it), and not required for normal use.

Review URL: http://codereview.chromium.org/507005