Changes to support ARM cross compilation.
Added a flag to allow verbose make-kpkg and a flag to specify the
cross compile tools.

Changed the default output_root to use ARCH in the path.

Modified determination of ARCH and changed ARCH=arm to the more correct
ARCH=armel. Also set KPKG_ARCH=arm for compliance with kernel makefile.

Disabled the setarch command when it fails and instead alias uname to
echo ARCH.

A call to compile for arm will look like:
./build_kernel.sh --config=path/to/config --cross_compile=arm-linux-gnueabi- \
[--verbose]

Review URL: http://codereview.chromium.org/501019