Adds Gtk theme.
Sorry Elliot, you're the only one that could review this. This is a
gtk theme implementation for Chrome OS. It's not complete, just having
enough for what Chrome uses. There will undoubtedly be tweaks going
forward, but this is a good start.

The name is pretty generic (theme), but that seems to fit in with the
name of the Chrome OS Window Manager (window_manager).

Review URL: http://codereview.chromium.org/502055