Change install_packages.sh to run from outside of the target rootfs and install debs into the target rootfs.
This will still run the maintainer scripts, but it is a step in
the right direction to being able to create a cross-compiled image.
This will allow us to experiment with disabling the maintainer
scripts and attempt to get that working. It also makes it easier
to modify the install_packages script since it now has access to
shflags and the source tree properly.

This CL also cleans up stuff doing in build_image.sh to bind mount
the trunk into the target rootfs, set up packages for install there,
etc.

The next step is to experiment with disabling maintainer scripts.
One possible way to do that would be to put a shim in place of dpkg
so that when apt calls dpkg it install a package we call dpkg-deb
with a simple unpack instead. This would allow us to continue to
use the good nice features that apt provides.

Review URL: http://codereview.chromium.org/517003