Add a helper script to be able to run "apt-get install" without running the maintainer scripts on installed packages.
I've experimented with using this in the install_packages.sh that
runs from outside the targetfs. It is able to extract packages and
make them appear to be installed (for the most part) without
running the maintainer scripts.

This will allow us to start figuring out what problems we will
have when we no longer run the maintainer scripts.

In the long run it would be nice to add an option to "dpkg" to
skip maintainer scripts.

Review URL: http://codereview.chromium.org/517004