Build dh-chromeos
dh-chromeos was originally intended for use once we switch to
chromiumos-build, but it turns out that it would be useful to be able to
use it beforehand to correct some cross-build problems. Accordingly,
give it a make_pkg.sh file and add it to build_platform_packages.sh.

Review URL: http://codereview.chromium.org/525030