Improve chromeos-metrics-daemon's build-dependencies
Add several necessary packages to Build-Depends for
chromeos-metrics-daemon, so that chromiumos-build can order builds
properly.

This is known not to be complete yet; in particular, it doesn't include
gflags, but we can't fix that yet because gflags doesn't have packaging
usable by chromiumos-build in the tree. Another CL will be forthcoming
once that's fixed, but this helps in the meantime.

Review URL: http://codereview.chromium.org/523085