This adds the option to not use clutter.
Now, if we supply --nouse_clutter to the window manager, we can render overview mode using straight OpenGL. It's not pretty yet, but it works.

Also, this adds a mock for the parts of gl and glX that we use, and expands the X connection interface so that we can do testing.

The test for no_clutter exists, and compiles, but doesn't test much yet.

Review URL: http://codereview.chromium.org/501030