Make drivers/lock.c build again
The original import of xscreensaver
(6316d91c27651ef356c7896bed3b1cfe953b1ef6) included changes to
driver/lock.c that removed the ignore_all_errors_ehandler function.
Unfortunately it's still referenced elsewhere in that file, so this
failed to build. I'm not sure why it was removed, but the simplest fix
appears to be to put the function back.

Review URL: http://codereview.chromium.org/523162