Add packages needed to build gtk+2.0 and pam
libatk1.0-doc, libcairo2-doc, libglib2.0-doc, and libpango1.0-doc are
currently required to build gtk+2.0 (which may be a bug, but if so then
this is a relatively harmless workaround for now); w3m is needed to
build pam. Add them all to dev.

Review URL: http://codereview.chromium.org/536020