Add optional chromiumos-build support to build_platform_packages
This makes it possible to attempt a build using chromiumos-build by
running './build_platform_packages.sh --new-build', with an -a option to
select the architecture (e.g. -a armel).

Right now, this uses --apt-source to build all build-dependencies as
well, since that usually ends up being what you want when
cross-compiling.  I'd appreciate feedback from buildbot-oriented folks
on whether this is an acceptable load in general.  For the meantime,
this is mainly for the convenience of people working on the cross-build
project.

Review URL: http://codereview.chromium.org/546042