Native build import script
Add a script to import all binary packages for a given source package
from ports.ubuntu.com, convert and install them using 'chromiumos-build
--convert', and write out a fake .changes file so that chromiumos-build
doesn't try to re-import them.  Use it like this:

  ./import_native_build.sh anthy libchewing

For the moment, there are a number of source packages that won't
cross-build successfully but that can be imported from native Ubuntu
builds, and this wraps up the process of doing so.

Review URL: http://codereview.chromium.org/543059