initial add of the pam_offline work
New PAM module to do offline authentication.  This works by using the given credentials to decrypt the user's master key.  If the decrypt is successful, the credentials pass, otherwise not.

This code will have to be kept in sync with the cryptohome scripts which initially create the master key(s).

Review URL: http://codereview.chromium.org/500034