Changes to support CFLAGS from env but still build with -fPIC.
We may want to provide a global CLFAGS, but the current SConstruct
pattern that we are using will take the global CFLAGS as a complete
replacement. Ideally what we want is to pullin the global CFLAGS but
still be able to specify "-fPIC -Wall -Werror -fno-exceptions", etc.

I'm not sure if this is the best way to do that, but it does solve
the short-term problem of things not building with -fPIC when CFLAGS
is present in the environment. Suggestions welcome.

Review URL: http://codereview.chromium.org/560028