This adds container attribute propagation.
Now when a container is transparent or transformed, the children of the container are affected as well.

Also, moved the layer visitor into the Tidy interface, and added support for testing whether an image has an alpha into is_opaque (and renamed to IsOpaque).

TEST=ran on netbook, tested overview mode.
BUG=none

Review URL: http://codereview.chromium.org/575005