Enable Chrome login!
Also, changes to enable logout and graceful (read: non-reboot) recovery if the session manager or X crash.

The first thing that needed to be done was to properly control the
upstart jobs that are part of the user's session, then make sure that
we clean up correctly after the session manager crashes/exits, and
then flip over to using the chrome login manager by default.

Review URL: http://codereview.chromium.org/573035