wm: Show panels immediately when mousing underneath them.
If the user moves the pointer down quickly to the bottom of
the screen, it's possible that it could end up below a
collapsed panel without us having received an enter event in
the panel's titlebar.  We show the panels immediately in
this case.

This also required some changes to pass the pointer position
through to event consumers for enter and leave events.

Review URL: http://codereview.chromium.org/579012