wm: Make event consumers register interest in things.
This replaces the previous model where events would be
offered to event consumers in an arbitrary order until
a consumer said that it had handled the event.

I'm still a bit torn between going with this or something
more general/flexible, where each consumer registers
interest in each specific type of event that it wants to
see.  I'm worried that that will result in a bunch more
code with little benefit -- do you have any thoughts about
this?

Other than the login stuff, the only other potential
additional event consumer is tidy (the library that we're
using instead of Clutter).  It only wants destroy, damage,
and configure notify events, none of which we send to
consumers right now, so I'm not sure if it's even worth
it making it an event consumer as opposed to just
hardcoding support for it in WindowManager.

BUG=none
TEST=included

Review URL: http://codereview.chromium.org/646020