Programmatic bash completion for Chromium OS build scripts.
If you source this script inside chroot, you would get <tab> completion for the --board= argument for some build scripts.

For example,
. bash_completion
./build_image --board=x<tab>
may complete to
./build_image --board=x86-generic
if x86-generic is the only available board starting with "x". If not, it will list available boards with the given prefix.

We could extend this to work with setup_board and other scripts. Also, it may be nice to extend it to work outside chroot.

Does it make sense to make this part of the regular dev chroot by default? If so, not sure what the best way to do that. Maybe install this as a hard-host-depends and source it from the chroot user .bashrc?

Review URL: http://codereview.chromium.org/652129